TDDN5-5 Фикс ошибки с определением типа long

+++ src/m3_ext/ui/base.py
@@ -226,7 +226,7 @@ class BaseExtComponent(object):
         elif isinstance(item, bool):
             res = str(item).lower()
 
-        elif isinstance(item, (int, float, decimal.Decimal, int)):
+        elif isinstance(item, (int, float, decimal.Decimal, long)):
             res = item
 
         elif isinstance(item, datetime.date):
